+++ test/server/spec/lib/args.spec.js
@@ -25,6 +25,7 @@ describe("lib/args", function () {
         builderrc: ".builderrc",
         help: false,
         logLevel: "info",
+        unlimitedBuffer: false,
         quiet: false,
         version: false
       });
@@ -39,6 +40,7 @@ describe("lib/args", function () {
         builderrc: dummyPath,
         help: false,
         logLevel: "info",
+        unlimitedBuffer: false,
         quiet: false,
         version: false
       });